Fix calls on the utils functions for getting features
These functions require the index of the primitive, but in splitddpg when we have one hardcoded primitives the "local" index is zero (the index of the network), not the global index of the primitive (0 for push target, 1, for push obstacle etc.). So the SplitDDPG should know which is the hardcoded primitive to call these functions correctly.